refactor(agent): isolate finalized SWM cleanup ownership
