+++ packages/agent/src/dkg-agent-base.ts
@@ -160,6 +160,7 @@ import {
 } from './auth/agent-delegation.js';
 import { SyncVerifyWorker } from './sync-verify-worker.js';
 import type { FinalizedSwmCleanupWorker } from './finalized-swm-cleanup-worker.js';
+import type { FinalizedSwmCleanupService } from './finalized-swm-cleanup-service.js';
 import { bindRandomSampling, type RandomSamplingDisabledReason, type RandomSamplingHandle, type RandomSamplingStatus } from './random-sampling-bind.js';
 import { connectToMultiaddr, ensurePeerConnected as ensurePeerConnectedAtom, primeCatchupConnections as primeCatchupConnectionsAtom } from './p2p/peer-connect.js';
 import { Messenger, type SloProtocolStats } from './p2p/messenger.js';
@@ -967,6 +968,7 @@ export class DKGAgentBase {
   protected swmCleanupTimer: ReturnType<typeof setInterval> | null = null;
   protected finalizedSwmCleanupTimer: ReturnType<typeof setInterval> | null = null;
   protected finalizedSwmCleanupWorker?: FinalizedSwmCleanupWorker;
+  protected finalizedSwmCleanupService?: FinalizedSwmCleanupService;
   /** Phase B — periodic chain-driven VM reconciliation sweep timer. */
   protected vmReconcileTimer: ReturnType<typeof setInterval> | null = null;
   /** Phase B — unified per-CG coalescing and node-wide admission policy. */
